+++ src/WinTool/CommandLine/CreateFileParameter.cs
@@ -1,4 +1,4 @@
-﻿using System.Text;
+using System.Text;
 
 namespace WinTool.CommandLine;
 
@@ -6,10 +6,8 @@ public class CreateFileParameter : ICommandLineParameter
 {
     public const string ParameterName = "/createFile";
     private const string PathSubParameter = "-path";
-    private const string SizeSubParameter = "-size";
 
     public string? FilePath { get; set; }
-    public long Size { get; set; }
 
     public void Parse(string arg)
     {
@@ -23,10 +21,6 @@ public class CreateFileParameter : ICommandLineParameter
             case PathSubParameter:
                 FilePath = values[1]; 
                 break;
-            case SizeSubParameter:
-                if (long.TryParse(values[1], out long size))
-                    Size = size;
-                break;
         }
     }
 
@@ -36,8 +30,6 @@ public class CreateFileParameter : ICommandLineParameter
 
         if (FilePath is not (null or []))
             sb.Append($" {PathSubParameter}=\"{FilePath}\"");
-        if (Size > 0)
-            sb.Append($" {SizeSubParameter}={Size}");
 
         return sb.ToString();
     }